+++ src/cpu/aarch64/matmul/brgemm_matmul_reorders.hpp
@@ -1,8 +1,6 @@
 /*******************************************************************************
 * Copyright 2022 Intel Corporation
 * Copyright 2024 FUJITSU LIMITED
-* Copyright 2025 Arm Ltd. and affiliates
-*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
@@ -27,12 +25,12 @@ namespace impl {
 namespace cpu {
 namespace aarch64 {
 
-struct brgemm_matmul_copy_reorder_t : public primitive_t {
+struct brgemm_matmul_matrix_B_reorder_t : public primitive_t {
     struct pd_t : public cpu_reorder_pd_t {
         using cpu_reorder_pd_t::cpu_reorder_pd_t;
 
-        DECLARE_COMMON_PD_T(
-                "brgemm_matmul_copy_reorder_t", brgemm_matmul_copy_reorder_t);
+        DECLARE_COMMON_PD_T("brgemm_matmul_matrix_B_reorder_t",
+                brgemm_matmul_matrix_B_reorder_t);
 
         // required to re-use brgemm matmul copy_b jit kernels
         matmul::brgemm_matmul_conf_t matmul_conf_for_reorder_;
@@ -49,7 +47,7 @@ struct brgemm_matmul_copy_reorder_t : public primitive_t {
         friend dnnl::impl::impl_list_item_t;
     };
 
-    brgemm_matmul_copy_reorder_t(const pd_t *apd) : primitive_t(apd) {}
+    brgemm_matmul_matrix_B_reorder_t(const pd_t *apd) : primitive_t(apd) {}
     status_t init(engine_t *engine) override {
         CHECK(matmul::create_brgemm_matmul_copy_b(
                 kernel_, &pd()->matmul_conf_for_reorder_));
@@ -66,9 +64,6 @@ private:
     const pd_t *pd() const { return (const pd_t *)primitive_t::pd().get(); }
     std::unique_ptr<matmul::jit_brgemm_matmul_copy_b_t> kernel_;
 };
-
-// Compatibility alias for code paths expecting matrix_B reorder on AArch64.
-using brgemm_matmul_matrix_B_reorder_t = brgemm_matmul_copy_reorder_t;
 
 } // namespace aarch64
 } // namespace cpu
